+++ src/main/java/org/tarak/pms/models/GoodsReceiveChallanItemId.java
@@ -8,13 +8,13 @@ public class GoodsReceiveChallanItemId implements Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private long goodsReceiveChallanId;
+	private int goodsReceiveChallanId;
 	private String finYear;
 	private int srNo;
-	public long getGoodsReceiveChallanId() {
+	public int getGoodsReceiveChallanId() {
 		return goodsReceiveChallanId;
 	}
-	public void setGoodsReceiveChallanId(long goodsReceiveChallanId) {
+	public void setGoodsReceiveChallanId(int goodsReceiveChallanId) {
 		this.goodsReceiveChallanId = goodsReceiveChallanId;
 	}
 	public String getFinYear() {
