+++ src/main/java/org/tarak/pms/models/PurchaseOrderId.java
@@ -8,12 +8,12 @@ public class PurchaseOrderId implements Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private long purchaseOrderId;
+	private int purchaseOrderId;
 	private String finYear;
-	public long getPurchaseOrderId() {
+	public int getPurchaseOrderId() {
 		return purchaseOrderId;
 	}
-	public void setPurchaseOrderId(long purchaseOrderId) {
+	public void setPurchaseOrderId(int purchaseOrderId) {
 		this.purchaseOrderId = purchaseOrderId;
 	}
 	public String getFinYear() {
