+++ src/main/java/org/tarak/pms/models/PurchaseOrderItemId.java
@@ -8,13 +8,13 @@ public class PurchaseOrderItemId implements Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private long purchaseOrderId;
+	private int purchaseOrderId;
 	private String finYear;
 	private int srNo;
-	public long getPurchaseOrderId() {
+	public int getPurchaseOrderId() {
 		return purchaseOrderId;
 	}
-	public void setPurchaseOrderId(long purchaseOrderId) {
+	public void setPurchaseOrderId(int purchaseOrderId) {
 		this.purchaseOrderId = purchaseOrderId;
 	}
 	public String getFinYear() {
